compat: Remove unused function
The compat function rpl_nf_conntrack_in() does not appear to be used
anywhere and emits warnings as such during builds < 4.10.

The patch passes Travis:

https://travis-ci.org/gvrose8192/ovs-experimental/builds/423097292

Remove it.

A backport of this patch to branch-2.10 seems to be required
to fix build errors when compiling against v3.16.54

https://travis-ci.org/openvswitch/ovs/jobs/597774225

Signed-off-by: Greg Rose <gvrose8192@gmail.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>
Signed-off-by: Simon Horman <simon.horman@netronome.com>
Reviewed-by: John Hurley <john.hurley@netronome.com>
---

This is an OVS backport targeted at upstream branch-2.10
(a bit like Linux stable)

I would like to make this public so I can push it into travis-ci
and proceed with unravelling any other brokenness in branch-2.10.