netdev-tc-offloads: mask ip_proto for pedit in ipv4/6 header and tcp/udp header
In the tc flower the pedit ipv4/6 header and tcp/udp header should
always mask the ip_proto. So it should set the mask before the
prio selection and it make the right prio selected in the non-support
multi_mask_per_prio kernel.

For case a rule with action dec_ttl. The flower now do't set the mask
of ip_proto and select the prio base on current masks. Then in the
tc_replace_filter it will mask the ip proto for pedit ip header action.

When the flow changs only with strip the action dec_ttl. The old tc rule
will be delete and the new one with the same prio (no ip_proto mask) and
install in the non multi_mask_per_prio kernel, it will be reject for
Invalid Arguement. because the current prio in kernel install rule with mask
ip_proto.

Signed-off-by: wenxu <wenxu@ucloud.cn>
Signed-off-by: Simon Horman <simon.horman@netronome.com>